fix: fix tabs error in console
