+++ apps/frontend/src/hooks/useDashBoard.ts
@@ -3,19 +3,23 @@ import { useContext, useMemo, useState } from "react";
 import { AVAILABLE_DASHBOARD_STATE } from "./types/dashBoardState";
 import { useRouter } from "next/router";
 import { AuthContext } from "@/context/AuthContext";
+import { Roles } from "@/components/Dashboard/StateLists";
 
 const useDashBoard = () => {
   const router = useRouter();
   const { user } = useContext(AuthContext);
-  const isGuide = useMemo(() => {
-    if (!user || !user.roles) {
-      return false;
+  const role: Roles = useMemo(() => {
+    if (user?.roles?.includes("GUIDE")) {
+      return Roles.GUIDE;
     }
-    return !!user.roles.includes("GUIDE");
+    if (user?.roles?.includes("ADMIN")) {
+      return Roles.ADMIN;
+    }
+    return Roles.USER;
   }, [user]);
   const [selectTab, setSelectTab] = useState<
     AVAILABLE_DASHBOARD_STATE | undefined
-  >(dashboardUrlMapper(router.asPath, isGuide));
+  >(dashboardUrlMapper(router.asPath, role));
 
   const onChange = (
     _: React.SyntheticEvent,
@@ -24,7 +28,7 @@ const useDashBoard = () => {
     setSelectTab(newValue);
   };
 
-  return { selectTab, onChange };
+  return { selectTab, role, onChange };
 };
 
 export default useDashBoard;
